finalise the navbar
