+++ src/component/FooterComponent/footerStyle.js
@@ -1,12 +1,12 @@
 import { makeStyles } from '@material-ui/core/styles';
 
 const useStyles = makeStyles(() => ({
-  root: {
-    position: 'relative',
-    bottom: 0,
-    left: 0,
-    width: '100%'
-  },
+  // root: {
+  //   position: 'relative',
+  //   bottom: 0,
+  //   left: 0,
+  //   width: '100%'
+  // },
 
   footerStyle: {
     display: 'flex',
